Added sproc mapping to table mapping Remove EFCore.Design dependency from EFCore.Specification.Tests
