+++ src/EFCore.Relational/Metadata/ITableMapping.cs
@@ -22,6 +22,21 @@ public interface ITableMapping : ITableMappingBase
     ///     Gets the properties mapped to columns on the target table.
     /// </summary>
     new IEnumerable<IColumnMapping> ColumnMappings { get; }
+    
+    /// <summary>
+    ///     Gets the corresponding insert stored procedure mapping if it exists.
+    /// </summary>
+    IStoredProcedureMapping? InsertStoredProcedureMapping { get; }
+    
+    /// <summary>
+    ///     Gets the corresponding insert stored procedure mapping if it exists.
+    /// </summary>
+    IStoredProcedureMapping? DeleteStoredProcedureMapping { get; }
+    
+    /// <summary>
+    ///     Gets the corresponding insert stored procedure mapping if it exists.
+    /// </summary>
+    IStoredProcedureMapping? UpdateStoredProcedureMapping { get; }
 
     /// <summary>
     ///     <para>
