+++ test/EFCore.Relational.Specification.Tests/Query/OwnedEntityQueryRelationalTestBase.cs
@@ -230,6 +230,97 @@ public abstract class OwnedEntityQueryRelationalTestBase : OwnedEntityQueryTestB
         }
     }
 
+    [ConditionalTheory]
+    [MemberData(nameof(IsAsyncData))]
+    public virtual async Task Owned_entity_with_all_null_properties_materializes_when_not_containing_another_owned_entity(bool async)
+    {
+        var contextFactory = await InitializeAsync<MyContext28247>(seed: c => c.Seed());
+
+        using var context = contextFactory.CreateContext();
+        var query = context.RotRutCases.OrderBy(e => e.Buyer);
+
+        var result = async
+            ? await query.ToListAsync()
+            : query.ToList();
+
+        Assert.Collection(result,
+            t =>
+            {
+                Assert.Equal("Buyer1", t.Buyer);
+                Assert.NotNull(t.Rot);
+                Assert.Equal(1, t.Rot.ServiceType);
+                Assert.Equal("1", t.Rot.ApartmentNo);
+                Assert.NotNull(t.Rut);
+                Assert.Equal(1, t.Rut.Value);
+            },
+            t =>
+            {
+                Assert.Equal("Buyer2", t.Buyer);
+                // Cannot verify owned entities here since they differ between relational/in-memory
+            });
+    }
+
+    protected class MyContext28247 : DbContext
+    {
+        public MyContext28247(DbContextOptions options)
+            : base(options)
+        {
+        }
+
+        public DbSet<RotRutCase> RotRutCases { get; set; }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            modelBuilder.Entity<RotRutCase>(b =>
+            {
+                b.ToTable("RotRutCases");
+
+                b.OwnsOne(e => e.Rot);
+                b.OwnsOne(e => e.Rut);
+            });
+        }
+
+        public void Seed()
+        {
+            Add(
+                new RotRutCase
+                {
+                    Buyer = "Buyer1",
+                    Rot = new Rot { ServiceType = 1, ApartmentNo = "1" },
+                    Rut = new Rut { Value = 1 }
+                });
+
+            Add(
+                new RotRutCase
+                {
+                    Buyer = "Buyer2",
+                    Rot = new Rot { ServiceType = null, ApartmentNo = null },
+                    Rut = new Rut { Value = null }
+                });
+
+            SaveChanges();
+        }
+    }
+
+    public class RotRutCase
+    {
+        public int Id { get; set; }
+        public string Buyer { get; set; }
+        public Rot Rot { get; set; }
+        public Rut Rut { get; set; }
+    }
+
+    public class Rot
+    {
+        public int? ServiceType { get; set; }
+        public string ApartmentNo { get; set; }
+    }
+
+    public class Rut
+    {
+        public int? Value { get; set; }
+    }
+
     protected override DbContextOptionsBuilder AddOptions(DbContextOptionsBuilder builder)
     {
         return base.AddOptions(builder).ConfigureWarnings(
