+++ test/EFCore.Relational.Tests/Metadata/Conventions/SequenceUniquificationConventionTest.cs
@@ -73,8 +73,8 @@ public class SequenceUniquificationConventionTest
     }
 
     private ProviderConventionSetBuilderDependencies CreateDependencies()
-        => RelationalTestHelpers.Instance.CreateContextServices().GetRequiredService<ProviderConventionSetBuilderDependencies>();
+        => FakeRelationalTestHelpers.Instance.CreateContextServices().GetRequiredService<ProviderConventionSetBuilderDependencies>();
 
     private RelationalConventionSetBuilderDependencies CreateRelationalDependencies()
-        => RelationalTestHelpers.Instance.CreateContextServices().GetRequiredService<RelationalConventionSetBuilderDependencies>();
+        => FakeRelationalTestHelpers.Instance.CreateContextServices().GetRequiredService<RelationalConventionSetBuilderDependencies>();
 }
