+++ test/EFCore.Relational.Tests/Storage/RelationalTypeMapperTest.cs
@@ -345,5 +345,5 @@ public class RelationalTypeMapperTest : RelationalTypeMapperTestBase
         => typeMappingSource.FindMapping(property);
 
     protected override ModelBuilder CreateModelBuilder(Action<ModelConfigurationBuilder> configure = null)
-        => RelationalTestHelpers.Instance.CreateConventionBuilder(configureModel: configure);
+        => FakeRelationalTestHelpers.Instance.CreateConventionBuilder(configureModel: configure);
 }
