+++ test/EFCore.Relational.Tests/Update/ModificationCommandComparerTest.cs
@@ -12,7 +12,7 @@ public class ModificationCommandComparerTest
     [ConditionalFact]
     public void Compare_returns_0_only_for_commands_that_are_equal()
     {
-        var modelBuilder = RelationalTestHelpers.Instance.CreateConventionBuilder();
+        var modelBuilder = FakeRelationalTestHelpers.Instance.CreateConventionBuilder();
         var entityType = modelBuilder.Model.AddEntityType(typeof(object));
         var key = entityType.AddProperty("Id", typeof(int));
         entityType.SetPrimaryKey(key);
@@ -163,7 +163,7 @@ public class ModificationCommandComparerTest
 
     private void Compare_returns_0_only_for_entries_that_have_same_key_values_generic<T>(T value1, T value2)
     {
-        var modelBuilder = RelationalTestHelpers.Instance.CreateConventionBuilder();
+        var modelBuilder = FakeRelationalTestHelpers.Instance.CreateConventionBuilder();
         var entityType = modelBuilder.Model.AddEntityType(typeof(object));
 
         var keyProperty = entityType.AddProperty("Id", typeof(T));
