+++ test/EFCore.Specification.Tests/QueryExpressionInterceptionTestBase.cs
@@ -1,8 +1,6 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using Microsoft.EntityFrameworkCore.Query.SqlExpressions;
-
 namespace Microsoft.EntityFrameworkCore;
 
 public abstract class QueryExpressionInterceptionTestBase : InterceptionTestBase
